+++ packages/harmony_log/lib/src/output/output.dart
@@ -30,6 +30,8 @@ abstract class LogOutput {
   /// redirect implementation
   ///
   /// redirect conditionally
+  ///
+  /// enabled by default
   const factory LogOutput.redirect({
     bool enabled,
     required LogOutput child,